requirements.txt
blinker==1.3
bottlenose==0.6.2
Flask==0.10.1
Flask-DebugToolbar==0.10.0
Flask-SQLAlchemy==2.0
httplib2==0.9.1
itsdangerous==0.24
Jinja2==2.7.3
lxml==3.4.4
MarkupSafe==0.23
python-amazon-product-api==0.2.8
python-amazon-simple-product-api==1.5.0
pytz==2015.2
six==1.9.0
SQLAlchemy==1.0.4
twilio==4.2.0
Werkzeug==0.10.4
